move classes into packages
