+++ src/main/java/org/folio/service/kafka/DIKafkaTopicService.java
@@ -1,6 +1,7 @@
-package org.folio.rest.impl.util;
+package org.folio.service.kafka;
 
 import org.folio.kafka.services.KafkaTopic;
+import org.folio.service.kafka.support.DataImportKafkaTopic;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Service;
@@ -13,11 +14,11 @@ public class DIKafkaTopicService {
   private Integer diInitNumPartitions;
 
   @Value("${di_raw_records_chunk_read.partitions}")
-  private Integer diRawRecordsChunkReadPartitions;
+  private Integer diRawRecordsChunkReadNumPartitions;
 
   public KafkaTopic[] createTopicObjects() {
     var initTopic = new DataImportKafkaTopic("DI_INITIALIZATION_STARTED", diInitNumPartitions);
-    var rawRecordsChunkReadTopic = new DataImportKafkaTopic("DI_RAW_RECORDS_CHUNK_READ", diRawRecordsChunkReadPartitions);
+    var rawRecordsChunkReadTopic = new DataImportKafkaTopic("DI_RAW_RECORDS_CHUNK_READ", diRawRecordsChunkReadNumPartitions);
     return new DataImportKafkaTopic[] {initTopic, rawRecordsChunkReadTopic};
   }
 }
