+++ src/test/java/org/folio/service/kafka/KafkaAdminClientServiceTest.java
@@ -27,8 +27,7 @@ import java.util.stream.Collectors;
 import org.apache.kafka.common.errors.TopicExistsException;
 import org.folio.kafka.services.KafkaAdminClientService;
 import org.folio.kafka.services.KafkaTopic;
-import org.folio.rest.impl.util.DIKafkaTopicService;
-import org.folio.rest.impl.util.DataImportKafkaTopic;
+import org.folio.service.kafka.support.DataImportKafkaTopic;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
